refactor: Money -> LottoMoney로 변경하고 검증 로직을 LottoMoney가 가지도록 변경
