+++ src/main/java/domain/LottoTickets.java
@@ -10,19 +10,16 @@ import java.util.stream.IntStream;
 public class LottoTickets {
     private final List<LottoTicket> lottoTickets;
     private final int selfPurchaseCount;
-    public static final String NOT_MULTIPLES_OF_PRICE_ERROR_MESSAGE
-            = String.format("금액을 %d의 배수로 입력해주세요.", LottoTicket.TICKET_PRICE);
 
     private LottoTickets(List<LottoTicket> lottoTickets, int selfPurchaseCount) {
         this.lottoTickets = lottoTickets;
         this.selfPurchaseCount = selfPurchaseCount;
     }
 
-    public static LottoTickets of(List<Set<Integer>> selfTicketNumbers, Money autoPurchaseMoney,
+    public static LottoTickets of(List<Set<Integer>> selfTicketNumbers, LottoMoney autoPurchaseLottoMoney,
                                   NumberGenerateStrategy numberGenerateStrategy) {
-        validatePurchaseMoney(autoPurchaseMoney);
         List<LottoTicket> lottoTickets = generateTicket(selfTicketNumbers);
-        int autoPurchaseCount = autoPurchaseMoney.getAmount() / LottoTicket.TICKET_PRICE;
+        int autoPurchaseCount = autoPurchaseLottoMoney.getAmount() / LottoTicket.TICKET_PRICE;
         lottoTickets.addAll(autoGenerateTickets(numberGenerateStrategy, autoPurchaseCount));
         return new LottoTickets(lottoTickets, selfTicketNumbers.size());
     }
@@ -38,16 +35,6 @@ public class LottoTickets {
         return IntStream.rangeClosed(1, autoPurchaseCount)
                 .mapToObj(index -> LottoTicket.from(numberGenerateStrategy.generateNumbers()))
                 .collect(Collectors.toList());
-    }
-
-    private static void validatePurchaseMoney(Money purchaseMoney) {
-        if (isMultiplesOfTicketPrice(purchaseMoney.getAmount())) {
-            throw new IllegalArgumentException(NOT_MULTIPLES_OF_PRICE_ERROR_MESSAGE);
-        }
-    }
-
-    private static boolean isMultiplesOfTicketPrice(int purchaseMoney) {
-        return purchaseMoney % LottoTicket.TICKET_PRICE != 0;
     }
 
     public List<LottoTicket> getTickets() {
